Add files via upload
In an undirected graph, edges are added over time. Write a program that, at certain moments, calculates the sum of the weights of edges in a connected component.

Input Format
The first line contains two integers, N and M (1 ≤ N, M ≤ 10^6), representing the number of vertices in the graph and the number of queries for adding edges.

Each of the next M lines contains a query.

Each query consists of either two or four numbers. The first number indicates the operation code.

If the first number is 1, it is followed by three numbers X, Y, W. This means an edge is added from vertex X to vertex Y with weight W (1 ≤ X, Y ≤ N, 1 ≤ W ≤ 10^3).
The graph may contain multiple edges (i.e., several edges between the same pair of vertices).
If the first number is 2, it is followed by exactly one number X. This means you need to answer the query: what is the sum of the weights of edges in the connected component containing vertex X (1 ≤ X ≤ N).
Output Format
For each operation with code 2, output the answer to the query on a separate line.

Example Input
text

Копировать
6 10
2 1
1 1 2 1
2 1
1 2 4 2
2 1
1 1 4 3
2 1
1 3 5 3
2 5
2 6
Example Output
text

Копировать
0
1
3
6
3
0